# api/announcements-openapi.yaml
openapi: 3.0.3
info:
  title: Announcements API
  version: 1.0.0
  description: REST API for managing announcements.
servers:
  - url: http://localhost:3001

tags:
  - name: Announcements

paths:
  /announcements:
    get:
      tags: [ Announcements ]
      operationId: getAnnouncements
      summary: List of all announcements
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
          description: Text search applied to title and content
        - in: query
          name: categories
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/CategoryCode"
          description: Filter announcements that contain at least one of selected categories
        - in: query
          name: sort
          required: false
          schema:
            $ref: "#/components/schemas/AnnouncementSortField"
          description: Sort field (default lastUpdate)
        - in: query
          name: order
          required: false
          schema:
            $ref: "#/components/schemas/SortOrder"
          description: Sort order (default desc)
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags: [ Announcements ]
      operationId: createAnnouncement
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewAnnouncement"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /announcements/{id}:
    get:
      tags: [ Announcements ]
      operationId: getAnnouncementById
      summary: Get announcement by id
      parameters:
        - $ref: "#/components/parameters/AnnouncementId"
      responses:
        "200":
          description: Announcement details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      tags: [ Announcements ]
      operationId: updateAnnouncement
      summary: Update announcement
      parameters:
        - $ref: "#/components/parameters/AnnouncementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnnouncementRequest"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      tags: [ Announcements ]
      operationId: deleteAnnouncement
      summary: Delete announcement
      parameters:
        - $ref: "#/components/parameters/AnnouncementId"
      responses:
        "200":
          description: Deleted announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /announcements/definitions:
    get:
      tags: [ Announcements ]
      operationId: getAnnouncementDefinitions
      summary: Get announcements definitions
      responses:
        "200":
          description: Definitions response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinitionsResponse"
        "500":
          $ref: "#/components/responses/ServerError"

components:
  parameters:
    AnnouncementId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: Announcement ID

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    CategoryCode:
      type: string
      enum:
        - CITY
        - COMMUNITY_EVENTS
        - CRIME_SAFETY
        - CULTURE
        - DISCOUNTS_BENEFITS
        - EMERGENCY
        - FOR_SENIORS
        - HEALTH
        - KIDS_FAMILY

    AnnouncementSortField:
      type: string
      enum:
        - title
        - publicationDate
        - lastUpdate

    SortOrder:
      type: string
      enum: [asc, desc]

    Announcement:
      type: object
      required: [id, title, content, publicationDate, lastUpdate, categories]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        publicationDate:
          type: string
          description: "MM/DD/YYYY HH:mm"
          example: "08/10/2025 08:55"
        lastUpdate:
          type: string
          description: "MM/DD/YYYY HH:mm"
          example: "08/10/2025 09:10"
        categories:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CategoryCode"

    AnnouncementListItem:
      type: object
      required: [id, title, publicationDate, lastUpdate, categories]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        publicationDate:
          type: string
          description: "MM/DD/YYYY HH:mm"
        lastUpdate:
          type: string
          description: "MM/DD/YYYY HH:mm"
        categories:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CategoryCode"

    AnnouncementListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AnnouncementListItem"
        total:
          type: integer
          minimum: 0
          description: Optional (if you decide to implement later)

    NewAnnouncement:
      type: object
      required: [title, content, publicationDate, categories]
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          maxLength: 5000
        publicationDate:
          type: string
          description: "MM/DD/YYYY HH:mm"
          pattern: '^(0[1-9]|1[0-2])\/(0[1-9]|[12]\d|3[01])\/\d{4}\s([01]\d|2[0-3]):[0-5]\d$'
        categories:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CategoryCode"

    UpdateAnnouncementRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          maxLength: 5000
        publicationDate:
          type: string
          description: "MM/DD/YYYY HH:mm"
          pattern: '^(0[1-9]|1[0-2])\/(0[1-9]|[12]\d|3[01])\/\d{4}\s([01]\d|2[0-3]):[0-5]\d$'
        categories:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CategoryCode"

    DefinitionsResponse:
      type: object
      additionalProperties: true

    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true